+++ CityBikes/LibraryTests/BikesNearbyTests.cs
@@ -30,13 +30,13 @@ namespace Library.Tests
         {
             var locations = AllBikesLocation.GetBikeLocations(context).ToArray();
 
-            List<Tuple<int, GPSLocation>> bikesExpected = new List<Tuple<int, GPSLocation>>() {
-                Tuple.Create(65530,new GPSLocation (4.12345678m,-4.12345678m)),
-                Tuple.Create(65531,new GPSLocation (2.12345678m,-2.12345678m)),
-                Tuple.Create(65532,new GPSLocation (6.12345678m,-13.12345678m)),
-                Tuple.Create(65533,new GPSLocation (8.12345678m,-2.12345678m)),
-                Tuple.Create(65534,new GPSLocation (2.12345678m,-10.12345678m)),
-                Tuple.Create(65535,new GPSLocation (5.12345678m,-1.12345678m))
+            List<Tuple<long, GPSLocation>> bikesExpected = new List<Tuple<long, GPSLocation>>() {
+                Tuple.Create(65530L,new GPSLocation (4.12345678m,-4.12345678m)),
+                Tuple.Create(65531L,new GPSLocation (2.12345678m,-2.12345678m)),
+                Tuple.Create(65532L,new GPSLocation (6.12345678m,-13.12345678m)),
+                Tuple.Create(65533L,new GPSLocation (8.12345678m,-2.12345678m)),
+                Tuple.Create(65534L,new GPSLocation (2.12345678m,-10.12345678m)),
+                Tuple.Create(65535L,new GPSLocation (5.12345678m,-1.12345678m))
             };
